Week-5 Assignment
Exercise 5.1 - Created 3 functions, and executed sequentially using both Async and Await.

Exercise 5.2 - Created function which accepts a string and returns a map where the keys are numbers and the values are the count of the vowels in the string.

Exercise 5.3 - Created a fn which accepts an array and returns true or false if that array contains a duplicate.

Exercise 5.4 - Created a simple Javascript function code for addition, subtraction, and multiplication of 2 numbers and write the corresponding Jest based tests for it.